+++ CHANGES.rst
@@ -24,6 +24,8 @@ Unreleased
   ``HistoricalRecords.context.request``) under some circumstances (gh-1188)
 - Made ``HistoryRequestMiddleware`` async-capable (gh-1209)
 - Fixed error when setting ``table_name`` with ``inherit=True`` (gh-1195)
+- Fixed ``FieldError`` when creating historical records for many-to-many fields with
+  ``to="self"`` (gh-1218)
 
 3.3.0 (2023-03-08)
 ------------------